DAS-1999: Input data's scalesize determines output dimensions.
output resolution before was determined simply by the scaleExtent divided by
the dimensions.  This meant that output images were similarly sized to input
data, but that the output resolution could be far different from the input data
source.

Now the resolution is determined from the input dataset and handled correctly
to determine the output dimensions. Including the degenerate case of a
projected data set with global coverage converting to a unprojected lat/lon grid.