DAS-2180: upgrade pillow with small work around changes
GDAL adds a NoData value to the png metadata when the colormap has 256 values
and only one of those is transparent: RGBA(0,0,0,0).

pillow 10.0.0 always returned a 256 length colormap from the quantized image
irrespective of the number of colors requested.

Pillow 10.x.0 changed this behavior to only return a palete of the unique
values in the quantized image.

By beginning with a 256-element all-black colortable and writing the truncated
one over it, we retain the previous behavior which allows for some images to
have a NoData Value in the PNG metadata.